requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
asn1crypto==1.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
natsort==8.4.0
packaging==24.2
pg8000==1.31.2
pluggy==1.5.0
python-dateutil==2.9.0.post0
requests==2.32.3
schedule==1.2.2
scramp==1.4.5
six==1.17.0
tabulate==0.9.0
urllib3==2.2.3

# For testing
pytest==8.3.4
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
coverage==7.3.2